Relocate MyArgs.
